Updated lramp function interface

+++ include/lsp-plug.in/dsp/common/pmath/lramp.h
@@ -28,21 +28,21 @@
  * @param dst destination vector
  * @param count number of elements
  */
-LSP_DSP_LIB_SYMBOL(void, lramp_set1, float *dst, float v1, float v2, uint32_t count);
+LSP_DSP_LIB_SYMBOL(void, lramp_set1, float *dst, float v1, float v2, size_t count);
 
 /** Apply linear ramping envelope: dst[i] = dst[i] * (v1 + ((v2-v1)*i) / count)
  *
  * @param dst destination vector
  * @param count number of elements
  */
-LSP_DSP_LIB_SYMBOL(void, lramp1, float *dst, float v1, float v2, uint32_t count);
+LSP_DSP_LIB_SYMBOL(void, lramp1, float *dst, float v1, float v2, size_t count);
 
 /** Apply linear ramping envelope: dst[i] = src[i] * (v1 + ((v2-v1)*i) / count)
  *
  * @param dst destination vector
  * @param count number of elements
  */
-LSP_DSP_LIB_SYMBOL(void, lramp2, float *dst, const float *src, float v1, float v2, uint32_t count);
+LSP_DSP_LIB_SYMBOL(void, lramp2, float *dst, const float *src, float v1, float v2, size_t count);
 
 
 /** Apply linear ramping envelope: dst[i] = dst[i] + src[i] * (v1 + ((v2-v1)*i) / count)
@@ -50,42 +50,42 @@ LSP_DSP_LIB_SYMBOL(void, lramp2, float *dst, const float *src, float v1, float v
  * @param dst destination vector
  * @param count number of elements
  */
-LSP_DSP_LIB_SYMBOL(void, lramp_add2, float *dst, const float *src, float v1, float v2, uint32_t count);
+LSP_DSP_LIB_SYMBOL(void, lramp_add2, float *dst, const float *src, float v1, float v2, size_t count);
 
 /** Apply linear ramping envelope: dst[i] = dst[i] - src[i] * (v1 + ((v2-v1)*i) / count)
  *
  * @param dst destination vector
  * @param count number of elements
  */
-LSP_DSP_LIB_SYMBOL(void, lramp_sub2, float *dst, const float *src, float v1, float v2, uint32_t count);
+LSP_DSP_LIB_SYMBOL(void, lramp_sub2, float *dst, const float *src, float v1, float v2, size_t count);
 
 /** Apply linear ramping envelope: dst[i] = src[i] * (v1 + ((v2-v1)*i) / count) - dst[i]
  *
  * @param dst destination vector
  * @param count number of elements
  */
-LSP_DSP_LIB_SYMBOL(void, lramp_rsub2, float *dst, const float *src, float v1, float v2, uint32_t count);
+LSP_DSP_LIB_SYMBOL(void, lramp_rsub2, float *dst, const float *src, float v1, float v2, size_t count);
 
 /** Apply linear ramping envelope: dst[i] = dst[i] * src[i] * (v1 + ((v2-v1)*i) / count)
  *
  * @param dst destination vector
  * @param count number of elements
  */
-LSP_DSP_LIB_SYMBOL(void, lramp_mul2, float *dst, const float *src, float v1, float v2, uint32_t count);
+LSP_DSP_LIB_SYMBOL(void, lramp_mul2, float *dst, const float *src, float v1, float v2, size_t count);
 
 /** Apply linear ramping envelope: dst[i] = dst[i] / (src[i] * (v1 + ((v2-v1)*i) / count))
  *
  * @param dst destination vector
  * @param count number of elements
  */
-LSP_DSP_LIB_SYMBOL(void, lramp_div2, float *dst, const float *src, float v1, float v2, uint32_t count);
+LSP_DSP_LIB_SYMBOL(void, lramp_div2, float *dst, const float *src, float v1, float v2, size_t count);
 
 /** Apply linear ramping envelope: dst[i] = (src[i] * (v1 + ((v2-v1)*i) / count)) / dst[i]
  *
  * @param dst destination vector
  * @param count number of elements
  */
-LSP_DSP_LIB_SYMBOL(void, lramp_rdiv2, float *dst, const float *src, float v1, float v2, uint32_t count);
+LSP_DSP_LIB_SYMBOL(void, lramp_rdiv2, float *dst, const float *src, float v1, float v2, size_t count);
 
 
 /** Apply linear ramping envelope: dst[i] = a[i] + b[i] * (v1 + ((v2-v1)*i) / count)
@@ -93,41 +93,41 @@ LSP_DSP_LIB_SYMBOL(void, lramp_rdiv2, float *dst, const float *src, float v1, fl
  * @param dst destination vector
  * @param count number of elements
  */
-LSP_DSP_LIB_SYMBOL(void, lramp_add3, float *dst, const float *a, const float *b, float v1, float v2, uint32_t count);
+LSP_DSP_LIB_SYMBOL(void, lramp_add3, float *dst, const float *a, const float *b, float v1, float v2, size_t count);
 
 /** Apply linear ramping envelope: dst[i] = a[i] - b[i] * (v1 + ((v2-v1)*i) / count)
  *
  * @param dst destination vector
  * @param count number of elements
  */
-LSP_DSP_LIB_SYMBOL(void, lramp_sub3, float *dst, const float *a, const float *b, float v1, float v2, uint32_t count);
+LSP_DSP_LIB_SYMBOL(void, lramp_sub3, float *dst, const float *a, const float *b, float v1, float v2, size_t count);
 
 /** Apply linear ramping envelope: dst[i] = b[i] * (v1 + ((v2-v1)*i) / count) - a[i]
  *
  * @param dst destination vector
  * @param count number of elements
  */
-LSP_DSP_LIB_SYMBOL(void, lramp_rsub3, float *dst, const float *a, const float *b, float v1, float v2, uint32_t count);
+LSP_DSP_LIB_SYMBOL(void, lramp_rsub3, float *dst, const float *a, const float *b, float v1, float v2, size_t count);
 
 /** Apply linear ramping envelope: dst[i] = a[i] * b[i] * (v1 + ((v2-v1)*i) / count)
  *
  * @param dst destination vector
  * @param count number of elements
  */
-LSP_DSP_LIB_SYMBOL(void, lramp_mul3, float *dst, const float *a, const float *b, float v1, float v2, uint32_t count);
+LSP_DSP_LIB_SYMBOL(void, lramp_mul3, float *dst, const float *a, const float *b, float v1, float v2, size_t count);
 
 /** Apply linear ramping envelope: dst[i] = a[i] / (b[i] * (v1 + ((v2-v1)*i) / count))
  *
  * @param dst destination vector
  * @param count number of elements
  */
-LSP_DSP_LIB_SYMBOL(void, lramp_div3, float *dst, const float *a, const float *b, float v1, float v2, uint32_t count);
+LSP_DSP_LIB_SYMBOL(void, lramp_div3, float *dst, const float *a, const float *b, float v1, float v2, size_t count);
 
 /** Apply linear ramping envelope: dst[i] = (b[i] * (v1 + ((v2-v1)*i) / count)) / a[i]
  *
  * @param dst destination vector
  * @param count number of elements
  */
-LSP_DSP_LIB_SYMBOL(void, lramp_rdiv3, float *dst, const float *a, const float *b, float v1, float v2, uint32_t count);
+LSP_DSP_LIB_SYMBOL(void, lramp_rdiv3, float *dst, const float *a, const float *b, float v1, float v2, size_t count);
 
 #endif /* LSP_PLUG_IN_DSP_COMMON_PMATH_LRAMP_H_ */
