+++ include/private/dsp/arch/x86/avx/pmath/lramp.h
@@ -39,7 +39,7 @@ namespace lsp
             };
         )
 
-        void lramp_set1(float *dst, float v1, float v2, uint32_t count)
+        void lramp_set1(float *dst, float v1, float v2, size_t count)
         {
             float delta = v2 - v1;
             if (delta == 0.0f)
@@ -181,7 +181,7 @@ namespace lsp
             __ASM_EMIT("8:")
 
 
-        void lramp1(float *dst, float v1, float v2, uint32_t count)
+        void lramp1(float *dst, float v1, float v2, size_t count)
         {
             float delta = v2 - v1;
             if (delta == 0.0f)
@@ -207,7 +207,7 @@ namespace lsp
             );
         }
 
-        void lramp2(float *dst, const float *src, float v1, float v2, uint32_t count)
+        void lramp2(float *dst, const float *src, float v1, float v2, size_t count)
         {
             float delta = v2 - v1;
             if (delta == 0.0f)
@@ -308,7 +308,7 @@ namespace lsp
             __ASM_EMIT("jge                 7b") \
             __ASM_EMIT("8:")
 
-        void lramp_add2(float *dst, const float *src, float v1, float v2, uint32_t count)
+        void lramp_add2(float *dst, const float *src, float v1, float v2, size_t count)
         {
             float delta = v2 - v1;
             if (delta == 0.0f)
@@ -334,7 +334,7 @@ namespace lsp
             );
         }
 
-        void lramp_sub2(float *dst, const float *src, float v1, float v2, uint32_t count)
+        void lramp_sub2(float *dst, const float *src, float v1, float v2, size_t count)
         {
             float delta = v2 - v1;
             if (delta == 0.0f)
@@ -360,7 +360,7 @@ namespace lsp
             );
         }
 
-        void lramp_rsub2(float *dst, const float *src, float v1, float v2, uint32_t count)
+        void lramp_rsub2(float *dst, const float *src, float v1, float v2, size_t count)
         {
             float delta = v2 - v1;
             if (delta == 0.0f)
@@ -386,7 +386,7 @@ namespace lsp
             );
         }
 
-        void lramp_mul2(float *dst, const float *src, float v1, float v2, uint32_t count)
+        void lramp_mul2(float *dst, const float *src, float v1, float v2, size_t count)
         {
             float delta = v2 - v1;
             if (delta == 0.0f)
@@ -412,7 +412,7 @@ namespace lsp
             );
         }
 
-        void lramp_div2(float *dst, const float *src, float v1, float v2, uint32_t count)
+        void lramp_div2(float *dst, const float *src, float v1, float v2, size_t count)
         {
             float delta = v2 - v1;
             if (delta == 0.0f)
@@ -438,7 +438,7 @@ namespace lsp
             );
         }
 
-        void lramp_rdiv2(float *dst, const float *src, float v1, float v2, uint32_t count)
+        void lramp_rdiv2(float *dst, const float *src, float v1, float v2, size_t count)
         {
             float delta = v2 - v1;
             if (delta == 0.0f)
@@ -464,7 +464,7 @@ namespace lsp
             );
         }
 
-        void lramp_add3(float *dst, const float *a, const float *b, float v1, float v2, uint32_t count)
+        void lramp_add3(float *dst, const float *a, const float *b, float v1, float v2, size_t count)
         {
             float delta = v2 - v1;
             if (delta == 0.0f)
@@ -492,7 +492,7 @@ namespace lsp
             );
         }
 
-        void lramp_sub3(float *dst, const float *a, const float *b, float v1, float v2, uint32_t count)
+        void lramp_sub3(float *dst, const float *a, const float *b, float v1, float v2, size_t count)
         {
             float delta = v2 - v1;
             if (delta == 0.0f)
@@ -520,7 +520,7 @@ namespace lsp
             );
         }
 
-        void lramp_rsub3(float *dst, const float *a, const float *b, float v1, float v2, uint32_t count)
+        void lramp_rsub3(float *dst, const float *a, const float *b, float v1, float v2, size_t count)
         {
             float delta = v2 - v1;
             if (delta == 0.0f)
@@ -548,7 +548,7 @@ namespace lsp
             );
         }
 
-        void lramp_mul3(float *dst, const float *a, const float *b, float v1, float v2, uint32_t count)
+        void lramp_mul3(float *dst, const float *a, const float *b, float v1, float v2, size_t count)
         {
             float delta = v2 - v1;
             if (delta == 0.0f)
@@ -576,7 +576,7 @@ namespace lsp
             );
         }
 
-        void lramp_div3(float *dst, const float *a, const float *b, float v1, float v2, uint32_t count)
+        void lramp_div3(float *dst, const float *a, const float *b, float v1, float v2, size_t count)
         {
             float delta = v2 - v1;
             if (delta == 0.0f)
@@ -604,7 +604,7 @@ namespace lsp
             );
         }
 
-        void lramp_rdiv3(float *dst, const float *a, const float *b, float v1, float v2, uint32_t count)
+        void lramp_rdiv3(float *dst, const float *a, const float *b, float v1, float v2, size_t count)
         {
             float delta = v2 - v1;
             if (delta == 0.0f)
