+++ include/private/dsp/arch/x86/sse/pmath/lramp.h
@@ -39,7 +39,7 @@ namespace lsp
             };
         )
 
-        void lramp_set1(float *dst, float v1, float v2, uint32_t count)
+        void lramp_set1(float *dst, float v1, float v2, size_t count)
         {
             float delta = v2 - v1;
             if (delta == 0.0f)
@@ -180,7 +180,7 @@ namespace lsp
             __ASM_EMIT("6:")
 
 
-        void lramp1(float *dst, float v1, float v2, uint32_t count)
+        void lramp1(float *dst, float v1, float v2, size_t count)
         {
             float delta = v2 - v1;
             if (delta == 0.0f)
@@ -206,7 +206,7 @@ namespace lsp
             );
         }
 
-        void lramp2(float *dst, const float *src, float v1, float v2, uint32_t count)
+        void lramp2(float *dst, const float *src, float v1, float v2, size_t count)
         {
             float delta = v2 - v1;
             if (delta == 0.0f)
@@ -307,7 +307,7 @@ namespace lsp
             \
             __ASM_EMIT("6:")
 
-        void lramp_add2(float *dst, const float *src, float v1, float v2, uint32_t count)
+        void lramp_add2(float *dst, const float *src, float v1, float v2, size_t count)
         {
             float delta = v2 - v1;
             if (delta == 0.0f)
@@ -333,7 +333,7 @@ namespace lsp
             );
         }
 
-        void lramp_sub2(float *dst, const float *src, float v1, float v2, uint32_t count)
+        void lramp_sub2(float *dst, const float *src, float v1, float v2, size_t count)
         {
             float delta = v2 - v1;
             if (delta == 0.0f)
@@ -359,7 +359,7 @@ namespace lsp
             );
         }
 
-        void lramp_rsub2(float *dst, const float *src, float v1, float v2, uint32_t count)
+        void lramp_rsub2(float *dst, const float *src, float v1, float v2, size_t count)
         {
             float delta = v2 - v1;
             if (delta == 0.0f)
@@ -385,7 +385,7 @@ namespace lsp
             );
         }
 
-        void lramp_mul2(float *dst, const float *src, float v1, float v2, uint32_t count)
+        void lramp_mul2(float *dst, const float *src, float v1, float v2, size_t count)
         {
             float delta = v2 - v1;
             if (delta == 0.0f)
@@ -411,7 +411,7 @@ namespace lsp
             );
         }
 
-        void lramp_div2(float *dst, const float *src, float v1, float v2, uint32_t count)
+        void lramp_div2(float *dst, const float *src, float v1, float v2, size_t count)
         {
             float delta = v2 - v1;
             if (delta == 0.0f)
@@ -437,7 +437,7 @@ namespace lsp
             );
         }
 
-        void lramp_rdiv2(float *dst, const float *src, float v1, float v2, uint32_t count)
+        void lramp_rdiv2(float *dst, const float *src, float v1, float v2, size_t count)
         {
             float delta = v2 - v1;
             if (delta == 0.0f)
@@ -463,7 +463,7 @@ namespace lsp
             );
         }
 
-        void lramp_add3(float *dst, const float *a, const float *b, float v1, float v2, uint32_t count)
+        void lramp_add3(float *dst, const float *a, const float *b, float v1, float v2, size_t count)
         {
             float delta = v2 - v1;
             if (delta == 0.0f)
@@ -491,7 +491,7 @@ namespace lsp
             );
         }
 
-        void lramp_sub3(float *dst, const float *a, const float *b, float v1, float v2, uint32_t count)
+        void lramp_sub3(float *dst, const float *a, const float *b, float v1, float v2, size_t count)
         {
             float delta = v2 - v1;
             if (delta == 0.0f)
@@ -519,7 +519,7 @@ namespace lsp
             );
         }
 
-        void lramp_rsub3(float *dst, const float *a, const float *b, float v1, float v2, uint32_t count)
+        void lramp_rsub3(float *dst, const float *a, const float *b, float v1, float v2, size_t count)
         {
             float delta = v2 - v1;
             if (delta == 0.0f)
@@ -547,7 +547,7 @@ namespace lsp
             );
         }
 
-        void lramp_mul3(float *dst, const float *a, const float *b, float v1, float v2, uint32_t count)
+        void lramp_mul3(float *dst, const float *a, const float *b, float v1, float v2, size_t count)
         {
             float delta = v2 - v1;
             if (delta == 0.0f)
@@ -575,7 +575,7 @@ namespace lsp
             );
         }
 
-        void lramp_div3(float *dst, const float *a, const float *b, float v1, float v2, uint32_t count)
+        void lramp_div3(float *dst, const float *a, const float *b, float v1, float v2, size_t count)
         {
             float delta = v2 - v1;
             if (delta == 0.0f)
@@ -603,7 +603,7 @@ namespace lsp
             );
         }
 
-        void lramp_rdiv3(float *dst, const float *a, const float *b, float v1, float v2, uint32_t count)
+        void lramp_rdiv3(float *dst, const float *a, const float *b, float v1, float v2, size_t count)
         {
             float delta = v2 - v1;
             if (delta == 0.0f)
