+++ src/test/ptest/pmath/lramp1.cpp
@@ -32,42 +32,42 @@ namespace lsp
 {
     namespace generic
     {
-        void lramp_set1(float *dst, float v1, float v2, uint32_t count);
-        void lramp1(float *dst, float v1, float v2, uint32_t count);
+        void lramp_set1(float *dst, float v1, float v2, size_t count);
+        void lramp1(float *dst, float v1, float v2, size_t count);
     }
 
     IF_ARCH_X86(
         namespace sse
         {
-            void lramp_set1(float *dst, float v1, float v2, uint32_t count);
-            void lramp1(float *dst, float v1, float v2, uint32_t count);
+            void lramp_set1(float *dst, float v1, float v2, size_t count);
+            void lramp1(float *dst, float v1, float v2, size_t count);
         }
 
         namespace avx
         {
-            void lramp_set1(float *dst, float v1, float v2, uint32_t count);
-            void lramp1(float *dst, float v1, float v2, uint32_t count);
+            void lramp_set1(float *dst, float v1, float v2, size_t count);
+            void lramp1(float *dst, float v1, float v2, size_t count);
         }
     )
 
     IF_ARCH_ARM(
         namespace neon_d32
         {
-//            void lramp_set1(float *dst, float v1, float v2, uint32_t count);
-//            void lramp1(float *dst, float v1, float v2, uint32_t count);
+//            void lramp_set1(float *dst, float v1, float v2, size_t count);
+//            void lramp1(float *dst, float v1, float v2, size_t count);
         }
     )
 
     IF_ARCH_AARCH64(
         namespace asimd
         {
-//            void lramp_set1(float *dst, float v1, float v2, uint32_t count);
-//            void lramp1(float *dst, float v1, float v2, uint32_t count);
+//            void lramp_set1(float *dst, float v1, float v2, size_t count);
+//            void lramp1(float *dst, float v1, float v2, size_t count);
         }
     )
 }
 
-typedef void (* lramp1_t)(float *dst, float v1, float v2, uint32_t count);
+typedef void (* lramp1_t)(float *dst, float v1, float v2, size_t count);
 
 //-----------------------------------------------------------------------------
 PTEST_BEGIN("dsp.pmath", lramp1, 5, 1000)
