+++ src/test/ptest/pmath/lramp2.cpp
@@ -32,67 +32,67 @@ namespace lsp
 {
     namespace generic
     {
-        void lramp2(float *dst, const float *src, float v1, float v2, uint32_t count);
-        void lramp_add2(float *dst, const float *src, float v1, float v2, uint32_t count);
-        void lramp_sub2(float *dst, const float *src, float v1, float v2, uint32_t count);
-        void lramp_rsub2(float *dst, const float *src, float v1, float v2, uint32_t count);
-        void lramp_mul2(float *dst, const float *src, float v1, float v2, uint32_t count);
-        void lramp_div2(float *dst, const float *src, float v1, float v2, uint32_t count);
-        void lramp_rdiv2(float *dst, const float *src, float v1, float v2, uint32_t count);
+        void lramp2(float *dst, const float *src, float v1, float v2, size_t count);
+        void lramp_add2(float *dst, const float *src, float v1, float v2, size_t count);
+        void lramp_sub2(float *dst, const float *src, float v1, float v2, size_t count);
+        void lramp_rsub2(float *dst, const float *src, float v1, float v2, size_t count);
+        void lramp_mul2(float *dst, const float *src, float v1, float v2, size_t count);
+        void lramp_div2(float *dst, const float *src, float v1, float v2, size_t count);
+        void lramp_rdiv2(float *dst, const float *src, float v1, float v2, size_t count);
     }
 
     IF_ARCH_X86(
         namespace sse
         {
-            void lramp2(float *dst, const float *src, float v1, float v2, uint32_t count);
-            void lramp_add2(float *dst, const float *src, float v1, float v2, uint32_t count);
-            void lramp_sub2(float *dst, const float *src, float v1, float v2, uint32_t count);
-            void lramp_rsub2(float *dst, const float *src, float v1, float v2, uint32_t count);
-            void lramp_mul2(float *dst, const float *src, float v1, float v2, uint32_t count);
-            void lramp_div2(float *dst, const float *src, float v1, float v2, uint32_t count);
-            void lramp_rdiv2(float *dst, const float *src, float v1, float v2, uint32_t count);
+            void lramp2(float *dst, const float *src, float v1, float v2, size_t count);
+            void lramp_add2(float *dst, const float *src, float v1, float v2, size_t count);
+            void lramp_sub2(float *dst, const float *src, float v1, float v2, size_t count);
+            void lramp_rsub2(float *dst, const float *src, float v1, float v2, size_t count);
+            void lramp_mul2(float *dst, const float *src, float v1, float v2, size_t count);
+            void lramp_div2(float *dst, const float *src, float v1, float v2, size_t count);
+            void lramp_rdiv2(float *dst, const float *src, float v1, float v2, size_t count);
         }
 
         namespace avx
         {
-            void lramp2(float *dst, const float *src, float v1, float v2, uint32_t count);
-            void lramp_add2(float *dst, const float *src, float v1, float v2, uint32_t count);
-            void lramp_sub2(float *dst, const float *src, float v1, float v2, uint32_t count);
-            void lramp_rsub2(float *dst, const float *src, float v1, float v2, uint32_t count);
-            void lramp_mul2(float *dst, const float *src, float v1, float v2, uint32_t count);
-            void lramp_div2(float *dst, const float *src, float v1, float v2, uint32_t count);
-            void lramp_rdiv2(float *dst, const float *src, float v1, float v2, uint32_t count);
+            void lramp2(float *dst, const float *src, float v1, float v2, size_t count);
+            void lramp_add2(float *dst, const float *src, float v1, float v2, size_t count);
+            void lramp_sub2(float *dst, const float *src, float v1, float v2, size_t count);
+            void lramp_rsub2(float *dst, const float *src, float v1, float v2, size_t count);
+            void lramp_mul2(float *dst, const float *src, float v1, float v2, size_t count);
+            void lramp_div2(float *dst, const float *src, float v1, float v2, size_t count);
+            void lramp_rdiv2(float *dst, const float *src, float v1, float v2, size_t count);
         }
     )
 
     IF_ARCH_ARM(
         namespace neon_d32
         {
-//            void lramp2(float *dst, const float *src, float v1, float v2, uint32_t count);
-//            void lramp_add2(float *dst, const float *src, float v1, float v2, uint32_t count);
-//            void lramp_sub2(float *dst, const float *src, float v1, float v2, uint32_t count);
-//            void lramp_rsub2(float *dst, const float *src, float v1, float v2, uint32_t count);
-//            void lramp_mul2(float *dst, const float *src, float v1, float v2, uint32_t count);
-//            void lramp_div2(float *dst, const float *src, float v1, float v2, uint32_t count);
-//            void lramp_rdiv2(float *dst, const float *src, float v1, float v2, uint32_t count);
+//            void lramp2(float *dst, const float *src, float v1, float v2, size_t count);
+//            void lramp_add2(float *dst, const float *src, float v1, float v2, size_t count);
+//            void lramp_sub2(float *dst, const float *src, float v1, float v2, size_t count);
+//            void lramp_rsub2(float *dst, const float *src, float v1, float v2, size_t count);
+//            void lramp_mul2(float *dst, const float *src, float v1, float v2, size_t count);
+//            void lramp_div2(float *dst, const float *src, float v1, float v2, size_t count);
+//            void lramp_rdiv2(float *dst, const float *src, float v1, float v2, size_t count);
         }
     )
 
     IF_ARCH_AARCH64(
         namespace asimd
         {
-//            void lramp2(float *dst, const float *src, float v1, float v2, uint32_t count);
-//            void lramp_add2(float *dst, const float *src, float v1, float v2, uint32_t count);
-//            void lramp_sub2(float *dst, const float *src, float v1, float v2, uint32_t count);
-//            void lramp_rsub2(float *dst, const float *src, float v1, float v2, uint32_t count);
-//            void lramp_mul2(float *dst, const float *src, float v1, float v2, uint32_t count);
-//            void lramp_div2(float *dst, const float *src, float v1, float v2, uint32_t count);
-//            void lramp_rdiv2(float *dst, const float *src, float v1, float v2, uint32_t count);
+//            void lramp2(float *dst, const float *src, float v1, float v2, size_t count);
+//            void lramp_add2(float *dst, const float *src, float v1, float v2, size_t count);
+//            void lramp_sub2(float *dst, const float *src, float v1, float v2, size_t count);
+//            void lramp_rsub2(float *dst, const float *src, float v1, float v2, size_t count);
+//            void lramp_mul2(float *dst, const float *src, float v1, float v2, size_t count);
+//            void lramp_div2(float *dst, const float *src, float v1, float v2, size_t count);
+//            void lramp_rdiv2(float *dst, const float *src, float v1, float v2, size_t count);
         }
     )
 }
 
-typedef void (* lramp2_t)(float *dst, const float *src, float v1, float v2, uint32_t count);
+typedef void (* lramp2_t)(float *dst, const float *src, float v1, float v2, size_t count);
 
 //-----------------------------------------------------------------------------
 PTEST_BEGIN("dsp.pmath", lramp2, 5, 1000)
